Update helloworld.java
Power test

+++ helloworld.java
@@ -8,9 +8,9 @@ public class helloworld
 		int d = a - b;
 		int e = a*b;
 		int f = a / b;
-		int a1 = 10;
-		int b1 = 5;
-		int c1 = math.pow(a1,b1);
+		double a1 = 10;
+		double b1 = 5;
+		double c1 = math.pow(a1,b1);
 		System.out.println("Subtraction of a and b = "+d);
 		System.out.println("Multiplication of a and b = "+e);
 		System.out.println("Division of a and b = "+f);
